+++ components/atoms/StyledHeadings/StyledHeadings.ts
@@ -1,0 +1,15 @@
+import styled from "styled-components";
+
+export const MainHeading = styled.h1`
+  font-size: 1.5rem;
+  margin: 1rem;
+
+  @media (min-width: 1024px) {
+    margin: 3rem 1rem;
+  }
+
+  @media (min-width: 768px) {
+    font-size: 2rem;
+    margin: 2.5rem 1rem;
+  }
+`;
