+++ components/organisms/CharactersListContainer/CharactersListContainer.tsx
@@ -1,19 +1,7 @@
-import styled from "styled-components";
+import { MainHeading } from "../../atoms/StyledHeadings/StyledHeadings";
 import { CharacterItems } from "../../../interfaces";
 import CharactersList from "../../molecules/CharactersList/CharactersList";
 import CharactersListPagination from "../../molecules/CharacterListPagination/CharacterListPagination";
-
-const MainHeading = styled.h1`
-  margin: 1rem;
-
-  @media (min-width: 1024px) {
-    margin: 3rem 1rem;
-  }
-
-  @media (min-width: 768px) {
-    margin: 2.5rem 1rem;
-  }
-`;
 
 type Props = {
   characters: CharacterItems;
