+++ components/templates/GlobalStyles/GlobalStyle.ts
